+++ src/modules/condominio/meu-condominio.module.ts
@@ -3,9 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Tenant } from '../../database/entities';
 import { MeuCondominioController } from './meu-condominio.controller';
 import { MeuCondominioService } from './meu-condominio.service';
+import { CepModule } from '../cep/cep.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Tenant])],
+  imports: [CepModule, TypeOrmModule.forFeature([Tenant])],
   controllers: [MeuCondominioController],
   providers: [MeuCondominioService],
 })